Improved Code Readability
* Refactor: Changed all comments description format
* Refactor: Add comments to CityServiceTest.java
* Refactor: Add CityServiceTest.java to pom.xml
* Add: testDecrementPartyVoteCount method
