+++ pom.xml
@@ -5,6 +5,7 @@
   <version>0.0.1-SNAPSHOT</version>
   
   <properties>
+	<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
   </properties>
@@ -26,7 +27,7 @@
             <version>3.0.0-M5</version>
             <configuration>
                 <includes>
-                    <include>**/*.java</include>
+                    <include>**/CityServiceTest.java</include>
                 </includes>
             </configuration>
        </plugin>
